% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan_config.R
\name{cmdstan_config}
\alias{cmdstan_config}
\alias{cmdstan_version}
\title{CmdStan configuration information}
\usage{
cmdstan_config()

cmdstan_version()
}
\value{
List with information on the current configuration of CmdStan.
  You can determine whether CmdStan+CmdStanPy was found using the `available`
  member (other members vary depending on whether `available` is `TRUE`
  or `FALSE`)
}
\description{
CmdStan configuration information
}
\keyword{internal}
