% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cmdstan.R
\name{install_cmdstan}
\alias{install_cmdstan}
\title{Convenience function to install CmdStan to your system
Note: assumes you have build tools installed (c++ etc)}
\usage{
install_cmdstan()
}
\description{
Convenience function to install CmdStan to your system
Note: assumes you have build tools installed (c++ etc)
}
