% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cmdstanpy.R
\name{install_cmdstanpy}
\alias{install_cmdstanpy}
\title{Install CmdStanPy and its dependencies}
\usage{
install_cmdstanpy(method = c("auto", "virtualenv", "conda"),
  conda = "auto", version = "default", envname = "r-cmdstanpy",
  extra_packages = NULL, restart_session = TRUE,
  conda_python_version = "3.6", ...)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows (as this isn't supported by TensorFlow). Note also
that since this command runs without privillege the "system" method is
available only on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
PATH and other conventional install locations).}

\item{version}{CmdStanPy version to install. Because CmdStanPy
  is not yet available on PyPI, this installs from github
  using git+git://github.com/stan-dev/cmdstanpy.

  Specify "default" to install from the "master" branch. Otherwise, provide a tag to the
  git commit/branch you wish to install.

  Alternatively, you can provide the full URL to an installer binary (e.g.
  for a nightly binary) or to an alternate git repo (e.g. your local fork).}

\item{envname}{Name of Python environment to install within}

\item{extra_packages}{Additional Python packages to install along with
TensorFlow.}

\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio).}

\item{conda_python_version}{the python version installed in the created conda
environment. Python 3.6 is installed by default.}

\item{...}{other arguments passed to [\code{reticulate::conda_install()}] or
[\code{reticulate::virtualenv_install()}].}
}
\description{
Install CmdStanPy and its dependencies
}
